refactor: densify research_planner_multi_slot pin + shared fixtures (#4310)
Extract multi-slot funding and decision-trace cases into sibling modules,
move Game/API helpers to shared support, and register the pin on the
residual-fat gated basename set (Slice C/D partial).
