+++ packages/colonizethis_ai/test/planning/research_planner_multi_slot_cases.dart
@@ -1,0 +1,9 @@
+// Case-library barrel for research-planner multi-slot pin (Refs #4310 Slice D).
+
+import 'research_planner_multi_slot_decision_trace_cases.dart';
+import 'research_planner_multi_slot_funding_cases.dart';
+
+void registerResearchPlannerMultiSlotCases() {
+  registerResearchPlannerMultiSlotFundingCases();
+  registerResearchPlannerMultiSlotDecisionTraceCases();
+}
