+++ tool/check_ai_residual_fat_pin_suite_size.dart
@@ -61,6 +61,7 @@ const Set<String> residualFatPinSuiteGatedBasenames = {
   'planning_invadable_owners_test.dart',
   'expand_phase_planner_sole_gp_war_helpers_test.dart',
   'expand_phase_planner_below_quota_peer_gp_peace_test.dart',
+  'research_planner_multi_slot_test.dart',
 };
 
 bool aiResidualFatPinSuiteSizePathInScope(String slashPath) {
